+++ src-tauri/plugins/android-webview/build.rs
@@ -1,4 +1,8 @@
-const COMMANDS: &[&str] = &[];
+const COMMANDS: &[&str] = &[
+    "save_backup_to_downloads",
+    "load_backup_from_downloads",
+    "set_haptic",
+];
 
 fn main() {
     tauri_plugin::Builder::new(COMMANDS)